Make --no_fwd_decls respect local forward decls
Before this, the --no_fwd_decls mode would force-remove existing forward
decls:

    class Foo;
    void Bar(const Foo*);

Add an additional rule that we avoid promoting uses of forward-
declarations in the same file to full uses, to retain forward-decls like
the above.

This makes the behavior of --no_fwd_decls simpler and more consistent:
optimize for fewer redeclarations instead of fewer includes. Whenever
there's a choice between a forward-decl and an include, choose the
include.

Add test coverage for more variants of included forward decls.