Fix wrong suggestions with --no_fwd_decls
This fixes 2 specific issues only happening with --no_fwd_decls:

(I) Removal of needed forward declaration from an associated header, when the
definiton is in the source file. (See #702)

(II) Replacing an include in an associated header file, containing a forward
declaration, with a local forward declaration, if the definiton can be found in
the header.